feat: enhance org admin dashboard with bug fixes and new widgets
- Fix published course count bug by setting is_published flag
- Fix active students tracking by updating last_login_at timestamp
- Add storage usage monitoring with KB/MB display for small sizes
- Add recent exams widget with click navigation
- Add popular courses widget ranked by student engagement
- Update storage schema to support decimal GB values (INTEGER to FLOAT)
- Reorganize dashboard layout with full-width activity chart
- Fix all syntax errors and restore missing cards

Backend changes:
- auth.py: Update last_login_at on login
- courses.py: Set is_published=True when publishing
- org.py: Add storage_usage, recent_exams, popular_courses to stats
- schemas.py: Change storage fields from int to float
- models.py: Update default storage_limit_gb to 50 GB

Frontend changes:
- page.tsx: Add new widgets and improve chart styling
- Weekly Activity chart now full width with gradient bars
- Storage shows adaptive units (KB/MB/GB)
- All cards properly styled with hover effects

+++ backend/app/schemas.py
@@ -99,8 +99,8 @@ class Organization(OrganizationBase):
     max_students: int
     max_teachers: int
     max_courses: int
-    storage_limit_gb: int
-    storage_used_gb: int
+    storage_limit_gb: float  # Changed to float to support decimal values
+    storage_used_gb: float   # Changed to float to support decimal values
     ai_credits_limit: int
     ai_credits_used: int
